Split governance tests from main TBTCSystemTests
Mostly because governance tests are quite extended at the moment. Also
restructure the governance test file to have the governance testing helper
function defined towards the tail end.
